requirements.txt
asgiref==3.5.2
Django==4.1.4
django-crispy-forms==1.14.0
django-registration-redux==2.11
Pillow==9.3.0
python-barcode==0.14.0
sqlparse==0.4.3
tzdata==2022.7
